+ mutex

+++ lib/batch/batch.go
@@ -17,15 +17,17 @@ func getOne(id int64) user {
 // according to the Semaphore concurrensy pattern lecture slide
 func getBatch(n int64, pool int64) (res []user) {
 	var mu sync.Mutex
-	res = make([]user, 0, int(n))
+	res = make([]user, 0, n)
 	var wg sync.WaitGroup
-	sem := make(chan struct{}, int(pool))
-	for i := 0; i < int(n); i++ {
+	sem := make(chan struct{}, pool)
+	for i := int64(0); i < n; i++ {
 		wg.Add(1)
-		sem <- struct{}{} // inside the go func in the video (wrong). less running goroutines
-		go func(i int) {  // it passes tests in bought position.
+		sem <- struct{}{} // right position. pool running goroutines
+		go func(i int64) {
+			// sem <- struct{}{}  // was inside the go func in the video (wrong).
+			// 100% running goroutine. pool - working
 			defer wg.Done()
-			user := getOne(int64(i))
+			user := getOne(i)
 			mu.Lock()
 			res = append(res, user)
 			mu.Unlock()
